a bit more trick rid nightly upcasting; tag 0.1.5
