--- rust-toolchain.toml
@@ -1,2 +1,0 @@
-[toolchain]
-channel = "nightly"
